+++ tests/TestMath.cpp
@@ -9,8 +9,8 @@
 
 BOOST_AUTO_TEST_CASE(testMath)
 {
-  BOOST_CHECK(exp(LogWeight(0).toDouble()) == RealWeight(0).toDouble());
-  BOOST_CHECK(log(RealWeight(0).toDouble()) == LogWeight(0).toDouble());
-  BOOST_CHECK(exp(LogWeight(1).toDouble()) == RealWeight(1).toDouble());
-  BOOST_CHECK(log(RealWeight(1).toDouble()) == LogWeight(1).toDouble());
+  BOOST_CHECK(exp(LogWeight(0)) == RealWeight(0));
+  BOOST_CHECK(log(RealWeight(0)) == LogWeight(0));
+  BOOST_CHECK(exp(LogWeight(1)) == RealWeight(1));
+  BOOST_CHECK(log(RealWeight(1)) == LogWeight(1));
 }
